BiSTNet-NTIRE2023/pip_requirements.txt
absl-py==1.0.0
addict==2.4.0
cachetools==4.2.4
charset-normalizer==2.0.12
click==8.0.4
cycler==0.11.0
dataclasses==0.8
decorator==4.4.2
easydict==1.9
einops==0.4.1
future==0.18.2
google-auth==2.6.0
google-auth-oauthlib==0.4.6
graphviz==0.19.1
grpcio==1.44.0
idna==3.3
imageio==2.15.0
imageio-ffmpeg==0.4.7
importlib-metadata==4.8.3
importlib-resources==5.4.0
kiwisolver==1.3.1
llvmlite==0.36.0
lmdb==1.3.0
logger==1.4
Markdown==3.3.6
matplotlib==3.3.4
moviepy==1.0.3
msgpack==1.0.4
networkx==2.5.1
numba==0.53.1
numpy==1.19.5
oauthlib==3.2.0
opencv-contrib-python==4.5.5.64
packaging==21.3
pandas==1.1.5
Pillow==8.4.0
prefetch-generator==1.0.1
proglog==0.1.10
protobuf==3.19.4
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.7
pypng==0.0.21
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.1.1
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.17.2
scipy==1.2.0
six==1.16.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tifffile==2020.9.3
timm==0.6.7
tqdm==4.63.0
typing_extensions==4.1.1
urllib3==1.26.8
Werkzeug==2.0.3
yacs==0.1.8
yapf==0.32.0
zipp==3.6.0
torchcontrib==0.0.2
